+++ lib/src/shader.cpp
@@ -12,16 +12,16 @@ OGAL::Shader::~Shader()
 
 void OGAL::Shader::load_shader(std::string shader_text, GLenum type)
 {
-    char const *shader_text_pointer = shader_text.c_str();
-
+    char const* shader_text_pointer = shader_text.c_str();
+    
     shader_id_ = glCreateShader(type);
     glShaderSource(shader_id_, 1, &shader_text_pointer, nullptr);
     glCompileShader(shader_id_);
-
+    
     glGetShaderiv(shader_id_, GL_COMPILE_STATUS, &result_);
     glGetShaderiv(shader_id_, GL_INFO_LOG_LENGTH, &log_lenght_);
     if (log_lenght_ > 0) {
-        std::vector<char> shader_error_message(log_lenght_ + 1);
+        std::vector<char> shader_error_message(log_lenght_ + (unsigned long long int) 1);
         glGetShaderInfoLog(shader_id_, log_lenght_, nullptr, &shader_error_message[0]);
         fprintf(stderr, "%s\n", &shader_error_message[0]);
     }
